Refactor configuration management and unit conversion utilities
- Removed the obsolete config.json file and refactored config.py to streamline configuration loading and validation.
- Introduced a new units.py module for energy and power unit conversions and validations.
- Eliminated the PGUtils class and replaced it with direct utility functions for power group management.
- Updated trace_recorders.py to remove unnecessary dependencies and simplify power group handling.
- Enhanced examples to utilize the new logging structure and improved energy consumption reporting.
- Added comprehensive unit tests for the new unit conversion functionalities and configuration validation.
- Removed deprecated example scripts and ensured all examples are aligned with the latest code structure.

+++ emt/power_groups/nvidia_gpu.py
@@ -113,9 +113,10 @@ class NvidiaGPU(PowerGroup):
         Returns:
             defaultdict: memory utilization of each zone
         """
-        used_mem_zones = defaultdict(int)
-        ps_mem_zones = defaultdict(int)
-        ps_util_zones = defaultdict(int)
+        used_mem_zones = defaultdict(float)
+        ps_mem_zones = defaultdict(float)
+        ps_util_zones = defaultdict(float)
+
         for zone, zone_handle in zip(self.zones, self._zones):
             try:
                 # get total used memory of a zone (zone: gpu 0,1,2 etc)
@@ -136,9 +137,9 @@ class NvidiaGPU(PowerGroup):
                     # memory usage of all filtered processes in a zone
                     zone_ps_memory += memory_used
                 # store values
-                used_mem_zones[zone] = zone_memory_used
+                used_mem_zones[zone] = float(zone_memory_used)
                 ps_mem_zones[zone] = zone_ps_memory
-                ps_util_zones[zone] = zone_ps_memory / zone_memory_used
+                ps_util_zones[zone] = zone_ps_memory / float(zone_memory_used)
             except pynvml.NVMLError as e:
                 logger.warning(f"Could not read utilizations due to NVML error: {e}")
         return used_mem_zones, ps_mem_zones, ps_util_zones
